Core: Huge rewrite and 2.5.1 opcodes
- Rewrite UpdateFields to Legacy system
- Change a lot of single line methods to use Expressions
- Remove whitespaces
- Remove deadcode
  - Achievement
  - Garrisons
  - Void Storage
  - Transmog
  - Toys
  - Azerite(Empowered)Item
  - Etc.
- Re-format files
- Use `out <type> <variable>` instead of forward declared out statements
- Fix some no value defined warnings (mostly packet related, will be rewritten)
- Remove all unused namespaces in files

+++ Source/Game/Chat/LanguageManager.cs
@@ -19,12 +19,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Game.DataStorage;
 using Framework.Constants;
-using Game.Spells;
-using Framework.Dynamic;
 using Framework.Collections;
 
 namespace Game.Chat
@@ -59,7 +55,7 @@ namespace Game.Chat
                     List<LanguageDesc> langsWithSkill = new();
                     foreach (var spellItr in spellsRange)
                         foreach (var skillPair in Global.SpellMgr.GetSkillLineAbilityMapBounds(spellItr.SpellId))
-                            langsWithSkill.Add(new LanguageDesc(spellItr.SpellId, (uint)skillPair.SkillLine));
+                            langsWithSkill.Add(new LanguageDesc(spellItr.SpellId, skillPair.SkillLine));
 
                     foreach (var langDesc in langsWithSkill)
                     {
